[Misc/Command] Reload command refactor
